+++ src/main/java/g1301_1400/s1354_construct_target_array_with_multiple_sums/Solution.java
@@ -21,7 +21,7 @@ public class Solution {
                 || target[maxIndex] % remainingSum == 0) {
             return false;
         }
-        target[maxIndex] %= remainingSum;
+        target[maxIndex] %= (int) remainingSum;
         return isPossible(target);
     }
 }
